Move all the number reducer functionality into the context.
