+++ core/src/main/kotlin/org/nwapw/abacus/plugin/standard/operator/OperatorAdd.kt
@@ -1,9 +1,9 @@
 package org.nwapw.abacus.plugin.standard.operator
 
 import org.nwapw.abacus.context.MutableEvaluationContext
-import org.nwapw.abacus.function.interfaces.NumberOperator
 import org.nwapw.abacus.function.OperatorAssociativity
 import org.nwapw.abacus.function.OperatorType
+import org.nwapw.abacus.function.interfaces.NumberOperator
 import org.nwapw.abacus.number.NumberInterface
 
 /**
