+++ core/src/test/java/org/nwapw/abacus/tests/RangeTests.java
@@ -6,8 +6,8 @@ import org.junit.Test;
 import org.nwapw.abacus.Abacus;
 import org.nwapw.abacus.config.Configuration;
 import org.nwapw.abacus.number.promotion.PromotionFunction;
-import org.nwapw.abacus.number.standard.NaiveNumber;
 import org.nwapw.abacus.number.range.NumberRange;
+import org.nwapw.abacus.number.standard.NaiveNumber;
 import org.nwapw.abacus.number.standard.PreciseNumber;
 import org.nwapw.abacus.plugin.standard.StandardPlugin;
